pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "phidata-agentfolio"
version = "0.1.0"
description = "AgentFolio integration for Phidata — agent identity, trust & reputation tools"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "phidata>=2.0.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://agentfolio.bot"
Repository = "https://github.com/0xbrainkid/phidata-agentfolio"
